+++ CHANGELOG.md
@@ -18,6 +18,7 @@ See [WIP.md](./WIP.md) for current work and bugs to fix
 
 ### Changed
 - webrtc - better error reporting
+- added Modules::stable for a default of all modules which should be used
 
 ### Fixed
 - enforce single tab and only one connection per node-ID